Move resty util funcs to client.go
