+++ pkg/artifactory/resource/configuration/configuration.go
@@ -2,14 +2,14 @@ package configuration
 
 import (
 	"github.com/go-resty/resty/v2"
-	"github.com/jfrog/terraform-provider-shared/util"
+	"github.com/jfrog/terraform-provider-shared/client"
 	"gopkg.in/ldap.v2"
 )
 
 func SendConfigurationPatch(content []byte, m interface{}) error {
 	_, err := m.(*resty.Client).R().SetBody(content).
 		SetHeader("Content-Type", "application/yaml").
-		AddRetryCondition(util.RetryOnMergeError).
+		AddRetryCondition(client.RetryOnMergeError).
 		Patch("artifactory/api/system/configuration")
 
 	return err
