+++ pkg/artifactory/resource/security/resource_artifactory_keypair.go
@@ -13,6 +13,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/jfrog/terraform-provider-artifactory/v6/pkg/artifactory/resource/repository"
+	"github.com/jfrog/terraform-provider-shared/client"
 	"github.com/jfrog/terraform-provider-shared/util"
 )
 
@@ -201,7 +202,7 @@ func createKeyPair(_ context.Context, d *schema.ResourceData, m interface{}) dia
 	keyPair, key, _ := unpackKeyPair(d)
 
 	_, err := m.(*resty.Client).R().
-		AddRetryCondition(util.RetryOnMergeError).
+		AddRetryCondition(client.RetryOnMergeError).
 		SetBody(keyPair).
 		Post(KeypairEndPoint)
 	if err != nil {
